refactor: update import paths for SolveResult and StepResult models
