+++ src/equation-solver-mobile-bff/presentation.application/domain/strategies/expression_solver.py
@@ -2,7 +2,7 @@ import re
 import math
 
 from domain.equations.errors import InvalidEquationError
-from domain.strategies.models_solver import SolveResult, StepResult
+from domain.strategies.models import SolveResult, StepResult
 from domain.strategies.strategy_solver import EquationSolverStrategy
