+++ src/equation-solver-mobile-bff/presentation.application/domain/strategies/factorization_solver.py
@@ -1,5 +1,5 @@
 from domain.equations.errors import InvalidEquationError
-from domain.strategies.models_solver import SolveResult, StepResult
+from domain.strategies.models import SolveResult, StepResult
 from domain.strategies.strategy_solver import EquationSolverStrategy
