+++ src/equation-solver-mobile-bff/presentation.application/domain/strategies/fractions_solver.py
@@ -2,7 +2,7 @@ import re
 from fractions import Fraction
 
 from domain.equations.errors import InvalidEquationError
-from domain.strategies.models_solver import SolveResult, StepResult
+from domain.strategies.models import SolveResult, StepResult
 from domain.strategies.strategy_solver import EquationSolverStrategy
