+++ src/equation-solver-mobile-bff/presentation.application/domain/strategies/inequality_solver.py
@@ -1,7 +1,7 @@
 import re
 
 from domain.equations.errors import InvalidEquationError
-from domain.strategies.models_solver import SolveResult, StepResult
+from domain.strategies.models import SolveResult, StepResult
 from domain.strategies.strategy_solver import EquationSolverStrategy
