+++ src/equation-solver-mobile-bff/presentation.application/domain/strategies/models/models_solver.py
