+++ src/equation-solver-mobile-bff/presentation.application/domain/strategies/strategy_solver.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from domain.strategies.models_solver import SolveResult
+from domain.strategies.models import SolveResult
 
 
 class EquationSolverStrategy(ABC):
